Menu Bar:   + can now open Properties View   + can now open Hierarchy View   + can now open Scene View   + can now open Project View   + can now open Console View
All buttons in the "View" dropdown on the Menu Bar are all functional.
If the window is not active, you can open a new one. You can only open
a new window if there is currently no active window for the one you chose.
Only 1 of each window can be active at a time.